ændret double til int i medlem

+++ src/Bruger/Formand.java
@@ -8,17 +8,17 @@ public class Formand {
     static Scanner scan = new Scanner(System.in);
 
     public static KonkurrenceMedlem opretMedlem(){
-        System.out.println("Navn: ");
+        System.out.println("Navn: " );
         String navn = scan.nextLine();
 
-        System.out.println("Alder: ");
+        System.out.println("Alder: " );
         int alder = scan.nextInt();
 
-        System.out.println("Aktiv Status: ");
+        System.out.println("Aktiv Status: " );
         boolean aktiv = scan.nextBoolean();
 
-        System.out.println("Medlemsnummer: ");
-        double medlemsnummer = scan.nextDouble();
+        System.out.println("Medlemsnummer: " );
+        int medlemsnummer = scan.nextInt();
 
         System.out.println("Restance: ");
         boolean restance = scan.nextBoolean();
